requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
kafka-python==2.0.2
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
requests==2.28.2